move components and lib out of src
