+++ new-dti-website/src/app/page.tsx
@@ -1,4 +1,4 @@
-import { Button } from '../components/ui/button';
+import { Button } from '../../components/ui/button';
 
 const Home = (): JSX.Element => (
   <div>
